💄 그룹 버튼 컴포넌트 분리
