+++ app/frontend/src/components/Group/index.tsx
@@ -3,10 +3,8 @@ import { ReactComponent as Crown } from '@/assets/icons/crown.svg';
 import { ReactComponent as Count } from '@/assets/icons/people.svg';
 import { vars } from '@/styles';
 
-import { useGroupJoinAndLeave } from './hooks/useGroupJoinLeave';
-import { useGroupModal } from './hooks/useGroupModal';
+import { GroupButton } from './GroupButton';
 import * as styles from './index.css';
-import { Button } from '../Button';
 
 const { grayscale200 } = vars.color;
 
@@ -17,14 +15,6 @@ type GroupProps = {
   name: string;
 };
 export function Group({ id, owned = false, name, joined = false }: GroupProps) {
-  const { handleLeave, handleJoin } = useGroupJoinAndLeave();
-  const { openLeaveModal, openJoinModal } = useGroupModal();
-
-  const onClickLeave = () =>
-    openLeaveModal({ onClickConfirm: () => handleLeave(id) });
-  const onClickJoin = () =>
-    openJoinModal({ onClickConfirm: () => handleJoin(id) });
-
   return (
     <div className={styles.container}>
       <div className={styles.titleWrapper}>
@@ -36,33 +26,7 @@ export function Group({ id, owned = false, name, joined = false }: GroupProps) {
             <span>200</span>
           </div>
         </div>
-        {owned && (
-          <Button type="button" theme="danger" shape="fill" size="medium">
-            그룹 삭제
-          </Button>
-        )}
-        {!owned &&
-          (joined ? (
-            <Button
-              type="button"
-              theme="danger"
-              shape="fill"
-              size="medium"
-              onClick={onClickLeave}
-            >
-              나가기
-            </Button>
-          ) : (
-            <Button
-              type="button"
-              theme="primary"
-              shape="fill"
-              size="medium"
-              onClick={onClickJoin}
-            >
-              참여하기
-            </Button>
-          ))}
+        <GroupButton id={id} owned={owned} joined={joined} />
       </div>
       {!owned && (
         <div className={styles.detail}>
